+++ examples/bar/config.py
@@ -8,7 +8,12 @@ from fabric.widgets.centerbox import CenterBox
 from fabric.system_tray.widgets import SystemTray
 from fabric.widgets.circularprogressbar import CircularProgressBar
 from fabric.widgets.wayland import WaylandWindow as Window
-from fabric.hyprland.widgets import Language, ActiveWindow, Workspaces, WorkspaceButton
+from fabric.hyprland.widgets import (
+    HyprlandLanguage,
+    HyprlandActiveWindow,
+    HyprlandWorkspaces,
+    WorkspaceButton,
+)
 from fabric.utils import FormattedString, get_relative_path, bulk_replace
 
 AUDIO_WIDGET = True
@@ -108,7 +113,7 @@ class StatusBar(Window):
             name="bar-inner",
             start_children=Box(
                 name="start-container",
-                children=Workspaces(
+                children=HyprlandWorkspaces(
                     name="workspaces",
                     spacing=4,
                     buttons_factory=lambda ws_id: WorkspaceButton(id=ws_id, label=None),
@@ -116,7 +121,7 @@ class StatusBar(Window):
             ),
             center_children=Box(
                 name="center-container",
-                children=ActiveWindow(name="hyprland-window"),
+                children=HyprlandActiveWindow(name="hyprland-window"),
             ),
             end_children=Box(
                 name="end-container",
@@ -126,7 +131,7 @@ class StatusBar(Window):
                     self.system_status,
                     SystemTray(name="system-tray", spacing=4),
                     DateTime(name="date-time"),
-                    Language(
+                    HyprlandLanguage(
                         name="hyprland-window",
                         formatter=FormattedString(
                             "{replace_lang(language)}",
