+++ fabric/core/widgets/__init__.py
@@ -1,0 +1,3 @@
+from .wm import WorkspaceButton, Workspaces, ActiveWindow, Language
+
+__all__ = ["WorkspaceButton", "Workspaces", "ActiveWindow", "Language"]
